hotfix: i can not describe what i just witnessed; bugs fixed; naming fixed; history fixed
